+++ src/inspect_ai/_display/rich.py
@@ -4,6 +4,7 @@ import datetime
 from dataclasses import dataclass
 from typing import Callable, Iterator, Set
 
+import rich
 from rich.console import Console, Group, RenderableType
 from rich.live import Live
 from rich.panel import Panel
@@ -34,6 +35,7 @@ from ._display import (
     TaskError,
     TaskProfile,
     TaskResult,
+    TaskScreen,
     TaskSuccess,
 )
 
@@ -77,13 +79,22 @@ class RichDisplay(Display):
 
     @override
     @contextlib.contextmanager
-    def live_task_status(self, total_tasks: int, parallel: bool) -> Iterator[None]:
+    def task_screen(self, total_tasks: int, parallel: bool) -> Iterator[TaskScreen]:
+        # reconfigure the default global console
+        use_color = is_running_in_vscode() and not is_running_in_jupyterlab()
+        rich.reconfigure(no_color=not use_color)
+
         self.total_tasks = total_tasks
         self.tasks = []
         self.progress_ui = rich_progress()
         self.parallel = parallel
         try:
-            with Live(None, console=rich_console(), auto_refresh=False) as live:
+            with Live(
+                None,
+                console=rich_console(),
+                transient=True,
+                auto_refresh=False,
+            ) as live:
                 # save reference to live
                 self.live = live
 
@@ -93,9 +104,10 @@ class RichDisplay(Display):
                 )
 
                 # yield
-                yield
+                yield RichTaskScreen(live)
 
                 # render task results
+                live.transient = False
                 live.update(tasks_results(self.tasks), refresh=True)
         finally:
             # clear tasks and progress
@@ -130,6 +142,7 @@ class RichDisplay(Display):
             and self.tasks
             and self.progress_ui is not None
             and self.live is not None
+            and self.live.is_started
         ):
             if self.parallel:
                 r = tasks_live_status(self.total_tasks, self.tasks, self.progress_ui)
@@ -138,6 +151,32 @@ class RichDisplay(Display):
             self.live.update(r, refresh=True)
 
         self.timer_handle = asyncio.get_event_loop().call_later(1, self._update_display)
+
+
+class RichTaskScreen(TaskScreen):
+    def __init__(self, live: Live) -> None:
+        self.live = live
+
+    @override
+    @contextlib.contextmanager
+    def input_screen(self, header: str | None = None) -> Iterator[Console]:
+        # clear live task status
+        self.live.update("", refresh=True)
+
+        # show cursor for input
+        self.live.console.show_cursor(True)
+
+        try:
+            # print header if requested
+            if header:
+                style = f"{rich_theme().meta} bold"
+                self.live.console.rule(f"[{style}]{header}[/{style}]", style=style)
+                self.live.console.print("")
+
+            # yield the console
+            yield self.live.console
+        finally:
+            self.live.console.show_cursor(False)
 
 
 class RichTaskDisplay(TaskDisplay):
@@ -574,13 +613,7 @@ def rich_theme() -> Theme:
 
 
 def rich_console() -> Console:
-    global _console
-    if _console is None:
-        # only use color in vscode (other terminals are too
-        # variable in their color contrast levels to rely on)
-        use_color = is_running_in_vscode() and not is_running_in_jupyterlab()
-        _console = Console(no_color=not use_color)
-    return _console
+    return rich.get_console()
 
 
 def rich_display() -> RichDisplay:
@@ -606,5 +639,4 @@ def rich_progress() -> RProgress:
 
 
 _theme: Theme | None = None
-_console: Console | None = None
 _display: RichDisplay | None = None