+++ src/inspect_ai/dataset/_dataset.py
@@ -15,6 +15,7 @@ from pydantic import BaseModel, Field
 from typing_extensions import override
 
 from inspect_ai.model import ChatMessage
+from inspect_ai.util import SandboxEnvironmentSpec
 
 if TYPE_CHECKING:
     from _typeshed import SupportsRichComparison
@@ -31,9 +32,11 @@ class Sample(BaseModel):
             or narrative text to be used by a model grader.
         id (int | str | None): Optional. Unique identifier for sample.
         metadata (dict[str,Any] | None): Optional. Arbitrary metadata associated with the sample.
+        sandbox (SandboxEnvironmentSpec | None): Optional. Sandbox environment
+          type and optional config file.
         files (dict[str, str] | None): Optional. Files that go along with the sample (copied to
           SandboxEnvironment). Files can be paths, inline text, or inline binary (base64 encoded data URL).
-        setup: (str | None): Optional. Setup script to run for sample (run
+        setup (str | None): Optional. Setup script to run for sample (run
           within default SandboxEnvironment).
     """
 
@@ -51,6 +54,9 @@ class Sample(BaseModel):
 
     metadata: dict[str, Any] | None = Field(default=None)
     """Arbitrary metadata associated with the sample."""
+
+    sandbox: SandboxEnvironmentSpec | None = Field(default=None)
+    """Sandbox environment type and optional config file."""
 
     files: dict[str, str] | None = Field(default=None)
     """Files that go along with the sample (copied to SandboxEnvironment)"""
@@ -159,7 +165,9 @@ class FieldSpec(BaseModel):
         choices (str): Optional. Name of field containing the list of answer choices.
         id (str): Optional. Unique identifier for the sample.
         metadata (list[str] | None): List of additional field names that should be read as metadata.
+        sandbox (str): Optional. Sandbox type along with optional config file
         files (str): Optional. Files that go along with the sample.
+        setup (str): Optional. Setup script to run for sample .
     """
 
     input: str = Field(default="input")
@@ -176,6 +184,9 @@ class FieldSpec(BaseModel):
 
     metadata: list[str] | None = Field(default=None)
     """List of additional field names that should be read as metadata."""
+
+    sandbox: str = Field(default="sandbox")
+    """Sandbox type along with optional config file."""
 
     files: str = Field(default="files")
     """Files that go along wtih the sample."""